backend/requirements.txt
fastapi
uvicorn
supabase
pydantic
pydantic-settings
python-dotenv
websockets
httpx
openai
pandas
qdrant-client
python-dateutil
faster-whisper
silero-vad
torch
sentence-transformers
PyPDF2
python-multipart
numpy
python-docx